+++ client/src/views/Login/Login.js
@@ -3,13 +3,14 @@ import Background from 'views/Background/Background';
 import useStyles from 'assets/jss/loginStyle';
 import backgroundImage from 'assets/img/background2.jpg'
 import gql from 'graphql-tag';
-import { ApolloClient, HttpLink, InMemoryCache } from 'apollo-boost';
-import { ApolloProvider,  Mutation } from 'react-apollo';
+
 import { Redirect } from 'react-router';
 import { useHistory } from "react-router-dom";
 import { withStyles } from '@material-ui/core/styles';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import { createHttpLink } from "apollo-link-http";
+import { useMutation } from '@apollo/react-hooks';
+
 const ColorCircularProgress = withStyles({
     root: {
       color: 'black',
@@ -42,15 +43,16 @@ export default function Login(props) {
     if (isAuth === null) { localStorage.setItem("isAuth", JSON.stringify(false)) }
     else if (isAuth) { history.push('/dashboard') }
 
-
-      
     const classes = useStyles();
     const [login, setLogin] = React.useState({
         name: '',
         password: '',
         token: '',
     });
+    const [loginMutation, { data }] = useMutation(LOGIN);
 
+    let inputName = '';
+    let inputPassword = '';
     useEffect(() => {
         console.log(login.token)
         if(!login.token || login.token === undefined) { return }
@@ -68,15 +70,21 @@ export default function Login(props) {
     const onHandleToken = (_token) => {
         props.passToken(_token);
     }
-    const onHandleMutate = (event, loginMutation) => {
+
+    return(
+    <Background  image={backgroundImage}>
+        <div className={classes.loginForm}>
+            <form onSubmit={(event) => {
+        console.log(inputName.value)
         event.preventDefault();
         loginMutation({
             variables: {
-                name: login.name,
-                password: login.password
+                name: inputName.value,
+                password: inputPassword.value
             }
         })
         .then((res) => {
+            console.log(inputName)
             const _token = res.data.login.token;
             setLogin({ ...login, token: _token });
             localStorage.setItem("isAuth", JSON.stringify(true));
@@ -85,43 +93,12 @@ export default function Login(props) {
         .catch((err)=> {
             alert('Your Account Is Not Valid!')
             console.log(err)
-        })
-    }
-
-    const headers = {
-        authorization: login.token ? `Bearer ${login.token}` : null
-    }
-    const httpLink = createHttpLink({
-    uri: 'http://localhost:4000',
-    credentials: 'same-origin'
-    })
-    const client = new ApolloClient({
-    link: httpLink,
-    headers : headers,
-    cache: new InMemoryCache()
-    })
-
-    return(
-    <ApolloProvider client={client}>
-    <Background  image={backgroundImage}>
-        <div className={classes.loginForm}>
-            <form>
+        })}}>
                 <p style={{color:'black',marginTop:'0px'}}>HYDROPONICS</p>
-                <input className={classes.login} placeholder="Name"  type="text" value={login.name} onChange={handleNameChange} />
-                <input className={classes.login} placeholder="Password" type="password" value={login.password} onChange={handlePasswordChange} />
+                <input ref={(node) => { inputName = node }} className={classes.login} placeholder="Name"  type="text" onChange={handleNameChange} />
+                <input ref={(node) => { inputPassword = node }} className={classes.login} placeholder="Password" type="password" value={login.password} onChange={handlePasswordChange} />
                 <div>
-                <Mutation mutation={LOGIN}>
-                {(loginMutation, { loading, error, data }) => {
-                    if (loading) 
-                        return <ColorCircularProgress size={15} thickness={1} /> ;
-                    return (
-                    <button className={classes.loginButton} 
-                            type="submit" 
-                            onClick={(event) => {
-                                onHandleMutate(event, loginMutation)}}>Log in</button>)
-                    }
-                }
-                </Mutation>
+                    <button className={classes.loginButton} type="submit" >Log in</button>
                 {/* <Query query={FEED}>
                 {({ data }) => {
                     return(<button className={classes.loginButton} 
@@ -135,10 +112,7 @@ export default function Login(props) {
                 </div>
             </form>
         </div>
-        {/* {localStorage.getItem("isAuth") ? <Redirect to="/dashboard" /> : null} */}
-    </Background>
-    </ApolloProvider>
-    
+    </Background>    
     )
     
 }
